refactor: project layout refactoring
Project layout refactoring according to golang standarts

Relates to #4
