+++ cmd/semantic-release/main.go
@@ -6,9 +6,9 @@ import (
 	"errors"
 	"flag"
 	"fmt"
-	"github.com/go-semantic-release/semantic-release"
-	"github.com/go-semantic-release/semantic-release/condition"
-	"github.com/go-semantic-release/semantic-release/update"
+	"github.com/go-semantic-release/semantic-release/pkg/semrel"
+	"github.com/go-semantic-release/semantic-release/pkg/condition"
+	"github.com/go-semantic-release/semantic-release/pkg/update"
 	"io/ioutil"
 	"log"
 	"os"
