+++ pkg/condition/travis.go
