+++ pkg/update/package_json_test.go
